Removed the default name of "" from new_vm, and the test for this value being "".
This makes sense. I now use argparse to ensure that the user enters a
name. Having the option to have a blank VM name, and creating a
template is a bad idea - we can then only do this once a day. Instead,
we should force the user to name their VM.