Added the capture-vm-snapshot script, which is used to turn a VM into a template.
Has been on husky0 for quite some time, as a matter of fact, so now is
about time that it be added to git.

When called, it takes the name of a VM as a parameter, then SSHs into
it, and wiats for the puppet process to die. When SSH returns, the VM is
shut down, and its VIF is removed. It is then turned into a template.