Reduced the value of {static,dynamic} minimum memory to be 256MiB.
Previously both max and min were set to 512MiB, because the freebie
version of XS doesn't support ballooning. Now we have the licensed
version, we can take RAM from machines if need be.

Now, normally this won't happen, and the machines will get all the RAM
that they have requested, and this patch won't make any difference to
the running on the VM.

However, there are 2 advantages of doing this
0) If we don't do it when we create the VM, it is hard to change these
settings when the VM is running. It involves shutting the VM down,
and fiddling in XenCentaur. We have been unable to do this for some VMs
we currently have.
1) By making VMs able to scale to very low (256MiB) amounts of memory,
it should be possible for us to run a complete (but reduced-performance)
service on single husky. Without this, we can only run on husky0, and
husky1. Given that we have hit plenty of issues where migrations, and
pools have not worked as expected, this may come in handy.