Remove a race condition in the ordering of SSH commands. The condition could result in a VM not getting the most up-to-date puppet config on creation. An error is thrown by cl-dtg-make-vm when this occurs.
Previously, the word looked like this:
-Run first SSH command that changed /etc/hostname, and ran vm-boot.
-Run another SSH command that pulls the latest version  of puppet and
applies it.

Applying puppet depends on both having the latest version, and having
the correct hostname, so we had these dependencies right.

However, the call to vm-boot, does a number of things. One of which is
deleting dom0's public key from /root/.ssh/authorized_keys.

Now, sometimes the above would work, because we spawn tasks using
nohups, with sleeps between. However sometimes it wouldn't.

Therefore, this patch ensures that we don't run vm-boot, before we make
the second connection into a guest. This ensures that we don't delete an
authorized key before the last connnection that needs it.