Specify the owner and deleteable by early on.
Previously, the setting of these values relied on a bunch of other steps
working, which may not have done. For instance, we had to RDNS a VM
before we set its owner.

This patch sets these details ASAP. Without it, it the cl-dtg-rm-vm
script will not run when a VM fails to create fully.