Add a third disk for swap.
At the moment we don't have a swap disk, which means that ballooning
breaks things when the vm should be writing out to swap.

The best way of providing swap is through an extra disk. If we use
partitions, resizing is annoying as we have to move around
partitions. With a third disk, we can just delete it and add a new one.